Tabs done and other things
There are now 2 tabs at the top of the items page, "all items" and "my
items". When the user looks at an item that they posted, they will see
the claim code. If an item is already claimed, the user will see that
instead of the claim button. Claimed items can still be seen by the user
that posted them under the "my items" tab.